refactor(lowlevel): further split into sub-packages
mainly for proper namespacing of ctor options
