+++ internal/lowlevel/encryptor/mod_test.go
@@ -1,4 +1,4 @@
-package lowlevel
+package encryptor
 
 import (
 	"testing"
