+++ internal/lowlevel/pkcs7/mod.go
@@ -1,6 +1,6 @@
-package lowlevel
+package pkcs7
 
-func pkcs7Pad(x []byte) []byte {
+func Pad(x []byte) []byte {
 	numPadBytes := 32 - len(x)%32
 	padByte := byte(numPadBytes)
 	tmp := make([]byte, len(x)+numPadBytes)
@@ -11,7 +11,7 @@ func pkcs7Pad(x []byte) []byte {
 	return tmp
 }
 
-func pkcs7Unpad(x []byte) []byte {
+func Unpad(x []byte) []byte {
 	// last byte is number of suffix bytes to remove
 	n := int(x[len(x)-1])
 	return x[:len(x)-n]
